[OPS-1559] Added parameters and separated  deployment commands
