+++ sc-app-deploy/sc-app-deploy.py
@@ -21,10 +21,8 @@ import logging
 import argparse
 from contextlib import redirect_stdout
 
-from sad_common.run_command import runCommand
 from sad_common.sad_logging import initLogging
-from sad_infra.application import Application
-from sad_infra.host import Host
+from sad_deploy.deploy_commands import deployDevelop, deployMaster, deployHotFix, deployFeature
 import sad_secrets.secret_helper
 
 def parseArguments():
@@ -78,86 +76,10 @@ def parseArguments():
     add_standard_args(feature_parser,
                       ('team-number', 'ticket-id'))
     feature_parser.set_defaults(func=deployFeature)
-    parser.add_argument('--debug', '-d', dest='debug', action='store_true', default=False),
-    parser.add_argument("-w", "--whatif", action='store_true', help = "If set no deploy operations are executed.")
 
     args = parser.parse_args()
     return args
     
-def logWhatIfHeader():
-    logging.info("====================================================")
-    logging.info("====================================================")
-    logging.info("====================================================")
-    logging.info("====== Simulation only! No backup performed. =======")
-    logging.info("====================================================")
-    logging.info("====================================================")
-    logging.info("====================================================")
-
-def deploy(application: Application, host: Host, decryptedSshKeyFile: str):
-    '''
-    Deploys the application to the given host.
-    '''
-    logging.info("Deploying '%s'..." % application.getSwarmServicename(host))
-
-    remoteUser = "travis"
-
-    sshCommand=['ssh']
-
-    # Disable known hosts checking.
-    sshOptions=['-o', 'StrictHostKeyChecking=no', '-o', 'UserKnownHostsFile=/dev/null']
-
-    if decryptedSshKeyFile != None:
-        # Use provided ssh key
-        sshIdentity=['-i', decryptedSshKeyFile]
-    else:
-        # Try pageant
-        sshIdentity=[]
-
-    # remote = <build user>@<hostname>.schul-cloud.dev
-    sshRemote=['%s@%s' % (remoteUser, host.getFQDN())]
-
-    # The remote command parameters.
-    # image = <imagename>:<imagetag> = <repository name>:<tag>
-    # service name = <hostname>_<applicationname short>. See 'docker service ls'
-    sshRemoteCommandParameters=[application.getImage(), application.getSwarmServicename(host)]
-
-    command = sshCommand + sshOptions + sshIdentity + sshRemote + sshRemoteCommandParameters
-    logging.info("Running command: '%s'" % ' '.join(command))
-
-    # Run docker service update
-    # Example call: ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i travisssh travis@hotfix6.schul-cloud.dev schulcloud/schulcloud-server:develop_latest hotfix6_server
-    # Example execution: /usr/bin/docker service update --force --image schulcloud/schulcloud-server:develop_latest hotfix6_server
-    runCommand(command)
-    #runCommand(['ls'])
-    logging.info("Deployment '%s' complete." % application.getSwarmServicename(host))
-
-    # TODO: Inform RocketChat
-
-def deployMaster():
-    pass
-
-def deployHotFix():
-    pass
-
-def deployFeature():
-    pass
-
-
-def deployDevelop(application: Application, host: Host, decrytedSshKeyFile: str):
-        decryptedSshKeyFile = None
-        if sad_secrets.secret_helper.isPassphraseSet():
-            decryptedSshKeyFile="travisssh"
-            sad_secrets.secret_helper.gpgDecrypt(decryptedSshKeyFile)
-        else:
-            logging.info("Passphrase not set in CI_GITHUB_TRAVISUSER_SWARMVM_KEY. Using ssh identity of the currently logged in user.")
-
-        # Deploy to the test host
-        deployHost = Host("hotfix6", "schul-cloud.dev")
-
-        deploy(Application("server", "schulcloud/schulcloud-server", "develop_latest"), deployHost, decryptedSshKeyFile)
-        deploy(Application("client", "schulcloud/schulcloud-client", "develop_latest"), deployHost, decryptedSshKeyFile)
-        deploy(Application("nuxtclient", "schulcloud/schulcloud-nuxt-client", "develop_latest"), deployHost, decryptedSshKeyFile)
-        deploy(Application("calendar", "schulcloud/schulcloud-calendar", "develop_latest"), deployHost, decryptedSshKeyFile)
 
 if __name__ == '__main__':
     try:
@@ -170,17 +92,13 @@ if __name__ == '__main__':
             print(os.environ['PATH'])
             sys.exit(1)
 
-        parsedArgs = parseArguments()
         initLogging()
-        if hasattr(parsedArgs, 'debug') and parsedArgs.debug:
-            rootLogger = logging.getLogger()
-            rootLogger.setLevel(logging.DEBUG)
-        logging.debug('Call arguments given: %s' % sys.argv[1:])
-        exit(0)
-        if hasattr(args, 'func'):
-            args.func(args)
+        parsedArgs = parseArguments()
+        logging.info('Call arguments given: %s' % sys.argv[1:])
+        if hasattr(parsedArgs, 'func'):
+            parsedArgs.func(parsedArgs)
         else:
-            logger.info("No command given, exiting ...")
+            logging.info("No command given, exiting ...")
         exit(0)
     except Exception as ex:
         logging.exception(ex)
